requirements.txt
asgiref==3.5.2
Django==6.0.3
djangorestframework==3.14.0
django-cors-headers==3.14.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.3
